refactor: extract durable tool payload
